cs688_retrieval/hw2/requirements.txt
tensorflow-gpu==1.4
matplotlib
tqdm
pillow
scipy
numpy
nanopq
imageio
animeface
pillow
scikit-image
